Edit security and use twilio

+++ src/main/java/santa/sistemas/tiendas_efe_springboot/repository/GenderRepository.java
@@ -3,5 +3,5 @@ package santa.sistemas.tiendas_efe_springboot.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import santa.sistemas.tiendas_efe_springboot.entity.Gender;
 
-public interface GenderRepository extends JpaRepository<Gender, Integer> {
+public interface GenderRepository extends JpaRepository<Gender, Long> {
 }
