+++ src/main/java/santa/sistemas/tiendas_efe_springboot/service/GenderService.java
@@ -2,6 +2,6 @@ package santa.sistemas.tiendas_efe_springboot.service;
 
 import santa.sistemas.tiendas_efe_springboot.entity.Gender;
 
-public interface GenderService extends iGenericService<Gender,Integer>{
+public interface GenderService extends iGenericService<Gender,Long>{
 
 }
